implementing the ACL on application

+++ app/Http/Livewire/Roles/Delete.php
@@ -3,6 +3,7 @@
 namespace App\Http\Livewire\Roles;
 
 use App\Models\Role;
+use Illuminate\Support\Facades\Gate;
 use Livewire\Component;
 use WireUi\Traits\Actions;
 
@@ -19,6 +20,10 @@ class Delete extends Component
 
     public function confirmDelete()
     {
+        if (!$this->isAuthorizable()) {
+            return;
+        }
+
         $this->dialog()->confirm([
             'title'       => 'Do you want to delete this role?',
             'description' => 'If you run this option, you will not be able to undo it.',
@@ -29,10 +34,25 @@ class Delete extends Component
 
     public function delete()
     {
+        if (!$this->isAuthorizable()) {
+            return;
+        }
+
         $this->role->forceDelete();
 
         $this->emitTo(All::class, 'roles::deleted');
 
         $this->notification()->success('Role deleted success');
+    }
+
+    private function isAuthorizable()
+    {
+        if (Gate::allows('forceDelete', $this->role)) {
+            return true;
+        }
+
+        $this->dialog()->error('You do not have authorization to perform this action');
+
+        return false;
     }
 }
