+++ app/Http/Livewire/Users/MoveToTrash.php
@@ -3,6 +3,7 @@
 namespace App\Http\Livewire\Users;
 
 use App\Models\User;
+use Illuminate\Support\Facades\Gate;
 use Livewire\Component;
 use WireUi\Traits\Actions;
 
@@ -19,6 +20,10 @@ class MoveToTrash extends Component
 
     public function confirmMoveToTrash()
     {
+        if (!$this->isAuthorizable()) {
+            return;
+        }
+
         $this->dialog()->confirm([
             'title'       => 'Are you Sure?',
             'description' => 'Do you want to move this user to trash?',
@@ -29,10 +34,25 @@ class MoveToTrash extends Component
 
     public function moveToTrash()
     {
+        if (!$this->isAuthorizable()) {
+            return;
+        }
+
         $this->user->delete();
 
         $this->emitTo(All::class, 'users::trashed');
 
         $this->notification()->success('User moved to trash success');
+    }
+
+    private function isAuthorizable()
+    {
+        if (Gate::allows('delete', $this->user)) {
+            return true;
+        }
+
+        $this->dialog()->error('You do not have authorization to perform this action');
+
+        return false;
     }
 }
